refactor: consolidate command directories into src/commands
Moved all command files from src/phantoms/commands/ and src/phantom/command/ to a single src/commands/ directory for better organization and maintainability.
