+++ src/commands/delete.test.ts
