+++ src/commands/list.test.ts
