+++ src/commands/version.test.ts
