+++ src/commands/version.ts
@@ -1,4 +1,4 @@
-import packageJson from "../../../package.json" with { type: "json" };
+import packageJson from "../../package.json" with { type: "json" };
 
 export function getVersion(): string {
   return packageJson.version;
