+++ src/commands/where.test.ts
